Add the weirdest image support I could come up with
Image files can be set as part of questions, they need to be inside a `images/` directory inside of the question directory (the one of that question set), they must have the same name as the question file, this also applies to file extension.
Images will resize to their full size, so it's not recommended to use big images, they are entirely optional as well.